Further improve logging
This is really just inlining the DoFailoverAsync method and adding some handy dandy logging.
